Create new `cos_integration` module
